@@ -1,11 +1,11 @@
 library flutter_slidable;
 
 export 'src/action_pane.dart';
-export 'src/action_pane_transitions.dart';
+export 'src/action_pane_motions.dart';
 export 'src/actions.dart';
 export 'src/controller.dart' show SlidableController, ActionPaneType;
 export 'src/dismissible_pane.dart';
-export 'src/dismissible_pane_transition.dart';
+export 'src/dismissible_pane_motions.dart';
 export 'src/notifications.dart'
     show SlidableNotification, SlidableNotificationListener;
 export 'src/slidable.dart';
