@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'slidable.dart';
 
+/// Signature for [CustomSlidableAction.onPressed].
 typedef SlidableActionCallback = void Function(BuildContext context);
 
 const int _kFlex = 1;
@@ -9,14 +10,14 @@ const Color _kBackgroundColor = Colors.white;
 const bool _kAutoClose = true;
 
 /// Represents an action of an [ActionPane].
-class SlidableAction extends StatelessWidget {
-  /// Creates a [SlidableAction].
+class CustomSlidableAction extends StatelessWidget {
+  /// Creates a [CustomSlidableAction].
   ///
-  /// The [flex], [backgroundColor], [foregroundColor] and [autoClose] must not
-  /// be null.
+  /// The [flex], [backgroundColor], [foregroundColor] and [autoClose] arguments
+  /// must not be null.
   ///
-  ///
-  const SlidableAction({
+  /// The [flex] argument must also be greater than 0.
+  const CustomSlidableAction({
     Key key,
     this.flex = _kFlex,
     this.backgroundColor = _kBackgroundColor,
@@ -30,11 +31,45 @@ class SlidableAction extends StatelessWidget {
         assert(child != null),
         super(key: key);
 
+  /// {@template slidable.actions.flex}
+  /// The flex factor to use for this child.
+  ///
+  /// The amount of space the child's can occupy in the main axis is
+  /// determined by dividing the free space according to the flex factors of the
+  /// other [CustomSlidableAction]s.
+  /// {@endtemplate}
   final int flex;
+
+  /// {@template slidable.actions.backgroundColor}
+  /// The background color of this action.
+  ///
+  /// Defaults to [Colors.white].
+  /// {@endtemplate}
   final Color backgroundColor;
+
+  /// {@template slidable.actions.foregroundColor}
+  /// The foreground color of this action.
+  ///
+  /// Defaults to [Colors.black] if [background]'s brightness is
+  /// [Brightness.light], or to [Colors.white] if [background]'s brightness is
+  /// [Brightness.dark].
+  /// {@endtemplate}
   final Color foregroundColor;
+
+  /// {@template slidable.actions.autoClose}
+  /// Whether the enclosing [Slidable] will be closed after [onPressed]
+  /// occurred.
+  /// {@endtemplate}
   final bool autoClose;
+
+  /// {@template slidable.actions.onPressed}
+  /// Called when the action is tapped or otherwise activated.
+  ///
+  /// If this callback is null, then the action will be disabled.
+  /// {@endtemplate}
   final SlidableActionCallback onPressed;
+
+  /// Typically the action's icon or label.
   final Widget child;
 
   @override
@@ -71,8 +106,8 @@ class SlidableAction extends StatelessWidget {
   }
 }
 
-class SlidableIconAction extends StatelessWidget {
-  const SlidableIconAction({
+class SlidableAction extends StatelessWidget {
+  const SlidableAction({
     Key key,
     this.flex = _kFlex,
     this.backgroundColor = _kBackgroundColor,
@@ -89,13 +124,25 @@ class SlidableIconAction extends StatelessWidget {
         assert(spacing != null),
         super(key: key);
 
+  /// {@macro slidable.actions.flex}
   final int flex;
+
+  /// {@macro slidable.actions.backgroundColor}
   final Color backgroundColor;
+
+  /// {@macro slidable.actions.foregroundColor}
   final Color foregroundColor;
+
+  /// {@macro slidable.actions.autoClose}
   final bool autoClose;
+
+  /// {@macro slidable.actions.onPressed}
   final SlidableActionCallback onPressed;
+
   final IconData icon;
+
   final double spacing;
+
   final String label;
 
   @override
@@ -136,7 +183,7 @@ class SlidableIconAction extends StatelessWidget {
             ],
           );
 
-    return SlidableAction(
+    return CustomSlidableAction(
       onPressed: onPressed,
       autoClose: autoClose,
       backgroundColor: backgroundColor,
